requirements.txt
boto3
botocore
pyspark==2.3.1
requests==2.18.4
psycopg2
python-dotenv
IPython
neo4j-driver==1.6.2
lxml